backend/requirements.txt
ansible==10.1.0
ansible-core==2.17.1
argcomplete==1.8.1
attrs==23.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.8
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.6.3
locust==2.32.5
Markdown==3.3.6
MarkupSafe==3.0.2
mkdocs==1.1.2
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
outcome==1.3.0.post0
packaging==21.3
paramiko==3.4.0
psutil==6.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycurl==7.44.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dotenv==0.21.1
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.2.0
requests==2.31.0
resolvelib==1.0.1
SecretStorage==3.3.1
selenium==4.22.0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
systemd-python==234
texttable==1.7.0
tomli==2.2.1
tornado==6.1
trio==0.25.1
trio-websocket==0.11.1
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.1
userpath==1.8.0
wadllib==1.3.6
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==1.0.0
zope.event==5.0
zope.interface==7.2
